+++ Solutions/SharePoint.Modernization/SharePointPnP.Modernization.Framework/Properties/AssemblyInfo.cs
@@ -32,7 +32,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.1.1808.0")]
-[assembly: AssemblyFileVersion("0.1.1808.0")]
+[assembly: AssemblyVersion("1.0.1810.0")]
+[assembly: AssemblyFileVersion("1.0.1810.0")]
 
 [assembly: InternalsVisibleTo("SharePointPnP.Modernization.Framework.Tests")]